Fix two bugs in the parallel runtime code.
One bug was caused when the master context, in MR_lc_finish() would release the
contexts used by each of the slots.  The release code attempts to save state
from the engine back into the context, which is necessary most of the time.
However, in this case it saved state from the engine running the master
context, into other contexts, so that when they where re-used they used an
invalid stack pointer.

Another bug was found in code with recursive parallel conjunctions.  Each
context structure contains a pointer to a code location, it is used as a value
for the instruction pointer when a context is resumed.  The
MR_join_and_continue operation for parallel conjunctions uses this resume to
ensure that the master context for a parallel conjunction is only resumed if it
has become blocked and is ready to be resumed.  However the field was never
cleared before and will always contain the same value parallel conjunctions are
nested as they will all have the same resume point.  This caused the master
context to be resumed before it had fully blocked, causing it to be resumed
with an invalid state.

A potential bug was found where a field should have been volatile to prevent
the compiler from caching its value when doing so would not be safe.

Widen a couple of critical sections as they didn't quite protect against some
race conditions.  This is another potential cause of bugs.

runtime/mercury_par_builtin.[ch]:
    Make the master_context field of the loop control structure volatile so
    that the compiler doesn't cache its value.

    Make the last worker to finish take a lock earlier, to ensure that the
    master context won't be left waiting forever.

    Add a comment explaining why a context must not be saved before calling
    MR_destroy_context().

    Improve debugging code to print out the value of the stack or parent stack
    pointer, depending on the code in question.

    Make the lock in MR_lc_finish() wider, so that the lock is held when the
    code checks to see if it should block.

runtime/mercury_context.c:
    MR_destroy_context() no longer saves the context before releasing it.

    MR_destroy_context() no longer sets the MR_ctxt_resume_owner_engine field
    of the context since it's not currently used.

    MR_join_and_continue(), the barrier for parallel conjunctions, how resets
    the resume code pointer of the master context when it switches to it.

runtime/mercury_context.h:
    Described the reason why the context must be saved before it is
    destroyed/released.

runtime/mercury_context.c:
runtime/mercury_engine.c:
    Call MR_save_context() before calling MR_destroy_context()