Fix one of the bugs that is preventing the trunk compiling with MSVC in the
Branches: main

Fix one of the bugs that is preventing the trunk compiling with MSVC in the
hlc.gc grade.  MSVC complains about the forward declaration of the DuArgLocn
array for the token/0 type in library/lexer.m having unknown size.  The array
size is not being printed in the forward declaration (as it should be) because
when we convert the RTTI to MLDS the initializer is set to be an initializer
for a struct, not an array.

compiler/rtti_to_mlds.m:
	Use the correct kind of initializer for DuArgLocn arrays.