The automatic parallelism analysis attempts to create conjunctions that can be parallelized by pushing lists of goals into other non-atomic goals, creating a suitable conjunction inside one or more branches of the non-atomic goal.  When this happens the costs and port counts of the pushed goals need to be updated so that they reflect what would happen if the code was re-ordered in this way.
There was a problem when the cost of the goal was re-calculated for recursive
calls.  The analysis re-calculated the percall cost from the total cost and the
new number of calls of the goal in this context.  This is incorrect, the
percall cost would not change simply because a call was called less often, the
percall cost would stay the same and the total number of calls would change.

deep_profiler/autopar_search_goals.m:
    Call goal_cost_change_calls which handles the update of a goal's cost when
    the number of calls to it changes.

deep_profiler/measurements.m:
    Create the goal_cost_change_calls function.