Add the --csharp-compiler-type and --output-csharp-compiler-type options.
Estimated hours taken: 4
Branches: main, 11.07

Add the --csharp-compiler-type and --output-csharp-compiler-type options.
This is an initial checkin of just the option so that we can bootstrap
a compiler which recognizes these options.

The options are needed, so that in conjunction with the --host-env-type
we can generate paths in the correct format for the csharp compiler.

compiler/mercury_compile.m:
compiler/options.m: