Fix a problem that was preventing the ROTD from building.
Branches: main

Fix a problem that was preventing the ROTD from building.  The recent change to
generate errors for abstract equivalence types that occur in type class
instances means that we now generate an error for the pred_id/0 type.
(It is an instance of the dependency_node/1 type class.)

compiler/hlds_pred.m:
	Put a notag wrapper around the body of the pred_id/0 type so
	that we can use it in type class instance without complaint.