Fix the rest of #255: the executable name of the the MS C# compiler clashes
Branches: main, 11.07

Fix the rest of #255: the executable name of the the MS C# compiler clashes
with that of the Chicken Scheme compiler.  This was leading to configure
incorrectly assuming the presence of the former when the latter was present.

The remainder of bug #255 relates to the configure script not building on
Mac OS X 10.4 -- I suspect the version of autoconf shipped with that system
is too ancient; I don't intend to do anything about it.  (This only applies
to generating the configure script.  The script itself, as provided in the
source distribution, should still run.)

m4/mercury.m4:
	Do not assume that the presence of an executable named csc indicates
	the presence of the MS C# compiler.  If csc is not the MS C# compiler,
	then attempt to use one of the others.

	Quote autoconf macro arguments in the Mercury's .NET configuration
	tests as recommended by the autoconf manual.