In the pass which adds a feature to tail call goals, ignore unifications of
Branches: main

In the pass which adds a feature to tail call goals, ignore unifications of
variables of dummy types.  Such unifications are treated by the code
generator as no-ops, so they should have no effect on whether a preceding
call is a tail call or not.

This pass affects the `--exec-trace-tail-rec' option (and, later,
the `--warn-non-tail-recursion' for low-level C grades).  For the former,
some potential tail calls would previously be inhibited in debugging grades.
For the latter, the compiler would produce false warnings about non-tail
recursive predicates.

compiler/mark_tail_calls.m:
	Bundle static information into a common structure.

	Make `find_maybe_output_args' ignore output variables of dummy types.

	Make `mark_tail_calls_in_goal' ignore unification goals involving
	a variable of a dummy type.