Suppress unused argument warnings in procedures generated by the loop-invariant
Branches: main

Suppress unused argument warnings in procedures generated by the loop-invariant
hoisting pass.

A variable which is determined to be invariant may become unused in the
auxiliary procedure after invariant goals are deleted, because the variable
only appears in invariant goals.  It is not that easy to detect such variables
(I tried) so we rely on the unused argument elimination pass to remove them,
and suppress the warnings.

compiler/unused_args.m:
	As above.

compiler/loop_inv.m:
	Add a comment about the unused arguments.