Begin work on integrating Mercury with Windows PowerShell.
Branches: main, 12.08 (probably)

Begin work on integrating Mercury with Windows PowerShell.

In the absence of Cygwin or MinGW/MSYS, using the Mercury compiler on Windows
has been rather troublesome, mostly due to the fact that cmd.exe is rubbish.
I've been experimenting with using PowerShell in place of cmd.exe recently
and my initial impressions are that it works much better.

This diff fixes a few issues that arise when COMSPEC is set to powershell.exe
rather than cmd.exe, in addition to adding powershell to the list of supported
environment types.  This is still work-in-progress.

compiler/globals.m:
	Add PowerShell to the list of environment types.

compiler/compile_target_code.m:
	With PowerShell we need to escape the `@' character: do so
	when generating command lines that use @files.

compiler/module_cmds.m:
	Unix and cmd.exe style redirection of standard input doesn't
	work on PowerShell.

	XXX add a couple of notes about things that could be improved.