Update and simplify the "tar" target in the top-level Mmakefile.
The main change here is to package up the test suite with the source
distribution rather than as a separate tarball.  The resulting combined tarball
is not significantly larger, so there is no good reason for not just
including the tests now.  TThe same rationale applies to not excluding
the benchmarks from the source distribution.

Mmakefile:
	Delete stuff in the tar target that was used for handling unused
	directories in the old CVS repository.

	Include the test suite in the main source distribution tarball
	rather than building a separate one for it.