Fix a problem with the valid/impure_detism test case.
GCC was emitting warnings about excess elements in structure initializers for
the above test.  The problem was that some #ifdefs were missing from the
initializers generated for MR_ProcStatic structures.  This meant that they did
not necessarily match the definition given for that structure in the runtime.

compiler/layout_out.m:
    Include the appropriate #ifdef guards for fields that only exist
    when MR_DEEP_PROFILING_COVERAGE is defined.