Avoid using sleep() in trace library on MinGW64.
The standard autoconf tests report that the sleep function is available on
MinGW64.  It isn't and we currently get a warning about an implicitly declared
function in the trace directory.   What is happening here is that the function
is not officially supported, and hence not in the library headers, but a symbol
corresponding to is exported from somewhere in the standard library.  (This is
what had happened last time something similar occurred on MinGW.)

The fix is to tell configure that sleep is not available on *mingw* and use
Sleep instead.

configure.ac:
     Don't attempt to use the function sleep on MinGW.