Build the documentation for the website in the main repository
Add support (actually there was some existing but unused support here
already) for building the documentation parts of the website within the doc/
directory of the main repository, rather than within the www repository
(which had to checkout the main repository).  This makes it easy to build
the documentation and upload it to the webserver, as the webserver has
limited resources.

doc/Mmakefile:
    As above.