Clarify licensing in READMEs.
Say that the GPL terms apply to the compiler and tools, not to the
"implementation".  As stated right after, the libraries and runtime are
distributed under the terms of the LGPL.

Don't say Mercury is free software "but" it is copyrighted.
There's nothing strange about it.

Fix spelling ("licence" is a noun).

.README.in:
bindist/bindist.README:
	As above.