Fix and enable the mfilterjavac option
Julien reported a bootstrap problem when mfilterjavac was enabled and a Java
grade was being installed but the installed compiler did not have an
mfilterjavac program.  When installing a Java grade the mfilterjavac program
in the workspace was not in the PATH.  This change ensures that mfilterjavac
is in the path and enables it's use in the compiler.

compiler/options.m:
    Enable the mfilterjavac option.

Mmakefile:
    Set the PATH in SUBDIR_MAKE and SUBSUBDIR_MAKE correctly so that the
    compiler can find mfilterjavac.

tools/bootcheck:
    Run 'mmake depend' in stage2/mfilterjavac in the same way that we do for
    the other directories.

mfilterjavac/.gitignore
    Ignore tags related files in this directory.