More fixes for float.float64_bits_string/1.
The implementation of float64_bits_string/1 does not work correctly on
platforms where sizeof(long) == 4.  This is the case for the i386 version of
NetBSD.  Also, make the code in the float module consistent with the fact the
runtime allows there to be no 64-bit integer type defined at all.

configure.ac:
runtime/mercury_conf.h.in:
	Define a new macro that expands to the integer length modifier for
	the MR_int_least64_t type.

library/float.m:
	Use the new macro to avoid a spot that was assuming that a long is
	a 64-bit quantity.  (TODO: we can probably now simplify some of the
	Windows specific code here, but I don't want to do that until I have
	tested it.)

	Have float64_bits_string/1 call MR_fatal_error() if MR_INT_LEAST64_TYPE
	is not defined.