Use a shallow copy for version_hash_table.copy
The code here was written ambiguously and one could not see at a glace if a
shallow or deep copy was being used, a shallow copy was always intended.
The deep copy of a version_array is unsupported on C backends and
unnecessary slow on any other backend.

librar/version_hash_table.m:
    Use the version_array.copy function (which is shallow) and make it clear
    by module qualifying the call.